kernel-doc parser: fix struct-member argument descriptions
It is very common to describe the members of a struct argument
with a dot-notation::

 * @layer:               ISDB per-layer data (only ISDB standard)
 * @layer.segment_count: Segment Count;
 * @layer.fec:           per layer code rate;
 * @layer.modulation:    per layer modulation;
 * @layer.interleaving:  per layer interleaving.

Without this patch, the kernel-doc parser prints something
buggy reST similar to::

  layer: per layer interleaving.

using the parameter name from the first line and the parameter
description from the last line.  This patch adds the ability to
handle kernel-doc comments using the dot-notation and output the
all descriptions:

  layer: ISDB per-layer data (only ISDB standard)
  layer.segment_count: Segment Count;
  layer.fec: per layer code rate;
  layer.modulation: per layer modulation;
  layer.interleaving: per layer interleaving.

As a side effect this patch also handles the "@args..."
parameter description of macros.

Signed-off-by: Markus Heiser <markus.heiser@darmarIT.de>