kernel-doc: add config value 'kernel_doc_mansect'
Global fallback for man section of kernel-doc directives. Set this value
if you want to create man pages for those kernel-doc directives, which
has not been set a ``:man-sect:`` value. The default is ``None``, which
means; do the opposite and create only man pages for those directives
which has been set the ``:man-sect:`` option (``None`` is what you
mostly want).

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>