kernel_doc mishandles declarations split into lines
This patches the error reported by Johannes Berg:

  https://www.mail-archive.com/linux-doc@vger.kernel.org/msg12410.html

When parsing a struct declaration which has fields splited into lines, the lines
has to be concatenate with an additional space. A one-liner of::

  struct something {
        struct foo
        bar;
        };

has to be::

  struct something {struct foo bar;};

Without the patch, the result missed the space between 'foo' and 'bar'.

The following link shows you the impact of this patch on the whole kernel source
tree:

  https://github.com/return42/sphkerneldoc/commit/56554e5f16c68d46d9c361e46540ce48ec1a9f27

Mostly spaces are added, but there are also some real bugs are fixed e.g.:

  https://github.com/return42/sphkerneldoc/commit/56554e5f16c68d46d9c361e46540ce48ec1a9f27#diff-523d26ea9da1bb7b97d2b763c75367e9

Signed-off-by: Markus Heiser <markus.heiser@darmarIT.de>