kernel-doc: assign only name in EXPORTED_SYMBOLS
When matching regular expression looking for exported symbol in macro
used to mark a symbol as exported is matched but also assigned. This
information is not used at all and this prevents some flexibility in the
form of regexp used if we wanted to replace it.

This modifies the regular expression used to find exported symbols to
assign only the exported symbol name.