kernel-doc: Don't mangle literal code blocks in comments
The origin patch comes from Jonathan Corbet [1] patching kernel-doc (perl)
parser of the Linux tree.

[1] https://www.mail-archive.com/linux-doc@vger.kernel.org/msg18046.html

--

It can be useful to put code snippets into kerneldoc comments; that can be
done with the "::" operator at the end of a line like this::

  if (desperate)
      run_in_circles();

kernel-doc currently fails to understand these literal blocks and applies
its normal markup to them, which is then treated as literal by sphinx.  The
result is unsightly markup instead of a useful code snippet.

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>