[fix] export option to kernel-doc directive
When `export` option is provided to kernel-doc directive without
option, the operation of collecting exported symbols by gather_context()
is skipped and it will trigger error soon after because of missing
exported symbols. It is however legitimate to use it without argument
since in can be configured globally on project configuration.

This ensures that exported symbols are populated when `export` option is
used in the directive no matter its value.